Kill _PORT_USE macro.
This trims down the number of sub-makes the port system runs quite a bit,
and makes some other fun stuff possible.

THIS KILLS THE pre-*/post-* SCRIPT STUFF
The functionality is unneeded, as it can be done with normal
pre-*/post-* targets, but it must be kept in mind when porting
Free/NetBSD ports.