A simple script to magically update-patches. It's a fair attempt at trying to match actual changed files with the patchdir, and invent new patch file names.
There are a few things it doesn't do yet, such as dealing with
patch files with multiple patches, or dealing with context diffs.

Just watch out when you use it for the first time on a port, verify
you don't lose or duplicate patches.

I've been using this to update large ports recently, and it seemed too
good not to give it to other porters.

Approved by turan@