Update to samba 2.0.7
This version was released on 25th April 2000 and includes the usual
slew of bug fixes and some new features, including Windows 2000
support.

One additional feature is the inclusion of the HTML of book "Using
Samba" which is made available through the SWAT interface. I am not
quite sure what the consensus is regarding HTML docs and the delivery
method for OpenBSD, so comments are, as usual, welcome.