Import of w3mir-1.0.9 From Jeff Bachtel <jeff@cepheid.org>, reviewed by naddy@
--
w3mir is a all purpose HTTP copying and mirroring tool. The main
focus of w3mir is to create and maintain a browseable copy of one,
or several, remote WWW site(s). Used to the max w3mir can retrieve
the contents of several related sites and leave the mirror browseable
via a local web server, or from a filesystem, such as directly from
a CDROM.

w3mir's goal is to be able to make useful mirrors of any reasonable
WWW site. It specifically preserves link integrity within the
mirrored documents as well as the integrety of links outside the
mirror, following redirects as needed. If you want it to. w3mir has
a powerful ``multi-scope'' mechanism enabling the user to make
mirrors of several related sites and have links between them refer
to the mirrored documents rather than the original site. w3mir has
several features directed at getting mirrors for CDROM burning and
handling of some not too often seen problems when mirroring.

w3mir supports HTML4, and has partial support for CSS, Java and
ActiveX.