Initial import of p5-Graphics-ColorNames-0.24
---
This module defines RGB values for common color names. The intention
is to (1) provide a common module that authors can use with other
modules to specify colors; and (2) free module authors from having
to "re-invent the wheel" whenever they decide to give the users the
option of specifying a color by name rather than RGB value.