Update to v3.29.12 - Install locale files. - Use the gettext port rather than building the included one. help from naddy@
Brief ChangeLog:
o Now it's possible to open a received URL with a browser directly;
controlled in the ~/.centericq/actions file.
o Cancelled messages are now being postponed for later editing.
o Substring search in history is implemented.
o QuickFind feature works in multiple contact selection mode as well
as in the contact list and is case insensitive.