Initial import of IPy-0.4
---
python module to handle IPv4 and IPv6 address

IPy is a module for handling IPv4 and IPv6 -Addresses and Networks
in a fashion similar to perl's Net::IP and friends. The IP class
allows a comfortable parsing and handling for most notations in use
for IPv4 and IPv6 Addresses and Networks.

WWW: http://c0re.jp:80/c0de/IPy/