Initial import of py-HappyDoc-2.0
---
python tool for generate Python API documents

HappyDoc is a tool for extracting documentation from Python source
code. It differs from other such applications by the fact that it
uses the parse tree for a module to derive the information and does
not import the module.

WWW: http://happydoc.sf.net/