Import pango 1.0.3. From DESCR:
Pango is a library for layout and rendering of text, with an emphasis
on internationalization. Pango can be used anywhere that text layout
is needed; however, most of the work on Pango-1.0 was done using the
GTK+ widget toolkit as a test platform. Pango forms the core of text
and font handling for GTK+-2.0.