Initial import of scli-0.2.12
The scli package was written to address the need for small and
efficient command line utilities to monitor and configure network
devices and host systems. The scli package is based on the SNMP
management protocol. It utilizes a MIB compiler called smidump to
generate C stub code.

Maintainer Margarida Sequeira <niness@devilness.org>