This ridiculous port installs the daemon 'ceptd' as setuid root and advises to be run as root via inetd, while it really only needs access to the modem It also needs INSTALL/DEINSTALL scripts to handle the configuration files, etc.  Try to bring it to a usable state, and advise the user to run it as follows:
xcept   stream  tcp     nowait  xcept   ${PREFIX}/libexec/ceptd ceptd

After creating a 'xcept' user belonging in group 'dialer'.  And, no
longer install this setuid root, which is really not needed.

XXX The user creation and the configuration files installation should
be automatic, but this will have to do for now and help the users.