lopster-1.0.1-dev5.12p1 update
lopster-1.0.1 release is eternally broken, both server_t and metaserver
changes have led to a broken port.  This version revives the port and will
be substituted for the release when it comes out.  I am normally against
any dev version going into the ports tree, but in this case the code has
tested more stable then the release (probably due to metaserver framework
changes) and it provides a working port as opposed to a broken one.  We
will rotate back to release when the author provides an update that solves
server side issues have been resolved as such.  I tried to back patch the
1.0.1 release, and it would require a substantial rewrite.

Features:

+ threaded connection implementation (connect to multiple networks at once)
+ framework for supporting other protocols although opennap is all that seems
  to work
+ patch-src_server_c is a back patch from cvs repo of napigator's changed url,
  without it, none of this comes together.

Tested on macppc (me), i386 (jolan), sparc64 (brad@) -- ok dev on this
approach rather then breaking port, brad@.