update to unrarsrc 3.2.1 (but it's called 3.20beta3 internally?!)
This update is based on an update by naddy@ done earlier this year.
I updated that to the current version and fixed sparc64 by disallowing
exceptions (otherwise core dumps would occur). I got some feedback from
wilfried@.

Main changes in the program are a seemingly total rewrite in C++ with
support for newer versions of the RAR algorithm (well, newer files that
didn't uncompress with the old unrar, uncompress with this).