Initial import of Maypole-Authentication-UserSessionCookie 1.1
This module allows Maypole applications to have the
concept of a user, and to track that user using cookies
and sessions.

It provides three methods to be inherited by a Maypole
class. The first is get_user, which tries to populate the
user slot of the Maypole request object.

from Sam Smith <S at msmith.net>