Initial import of Email-MIME-1.82
This is an extension of the Email::Simple module, to handle
MIME encoded messages. It takes a message as a string,
splits it up into its constituent parts, and allows you
access to various parts of the message. Headers are decoded
from MIME encoding.

from Sam Smith <s at msmith.net>