Initial import of Tie-IxHash-1.21.
This Perl module implements Perl hashes that preserve the order in which
the hash elements were added. The order is not affected when values
corresponding to existing keys in the IxHash are changed. The elements
can also be set to any arbitrary supplied order. The familiar perl array
operations can also be performed on the IxHash.