import dwdiff 1.2
dwdiff is a front-end for the diff program that operates at the word
level instead of the line level. It is different from wdiff in that it
allows the user to specify what should be considered whitespace, and in
that it takes an optional list of characters that should be considered
delimiters. Delimiters are single characters that are treated as if they
are words, even when there is no whitespace separating them from
preceding words or delimiters. dwdiff is mostly commandline compatible
with wdiff. Only the --autopager, --terminal and --avoid-wraps options
are not supported.

feedback & ok steven