Initial import of jdk 1.7 (a.k.a. OpenJDK 7) work in progress. So far Sun has released the hotspot portion of 1.7 under GPLv2. j2se and deploy (plugin & webstart) will come later. No packages yet since j2se is copied in from 1.5.
Note: hotspot for BSD in previous versions of the jdk were based on a
mixture of Solaris code and Linux code (for amd64). In this version I have
selected Linux as the base and only ported of small parts of the Solaris
code as needed. Using Linux code as the base has significantly reduced the
code divergence for BSD support and should make maintaining the port
easier.

okay ian@, go for it espie@