Import gnustep-back-0.12.0
This is the back-end component for the GNUstep GUI Library.  The
implementation of the GNUstep GUI Library is designed in two parts.  The
first part is the front-end component which is independent of platform
and display system.  This front-end is combined with a back-end
component which handles all of the display system dependent such as
specific calls to the X Window System.  This design allows the GNUstep
applications to have the "look and feel" of the underlying display
system without any changes to the application, and the library can be
easily ported to other display systems.
The GNUstep GUI Backend is for platforms using the X-Window System or
Window's Systems. It works via a DPS emulation engine to emulate the
DPS functions required by the front-end system.