import py-paste-deploy-1.3
Paste Deployment is a system for finding and configuring WSGI
applications and servers. For WSGI application consumers it provides
a single, simple function (loadapp) for loading a WSGI application
from a configuration file or a Python Egg. For WSGI application
providers it only asks for a single, simple entry point to your
application, so that application users don't need to be exposed to
the implementation details of your application.

ok jasper@