import libsigsegv 2.4
This is a library for handling page faults in user mode. A page fault
occurs when a program tries to access to a region of memory that is
currently not available. Catching and handling a page fault is a useful
technique for implementing:

    * pageable virtual memory,
    * memory-mapped access to persistent databases,
    * generational garbage collectors,
    * stack overflow handlers,
    * distributed shared memory,

based on a submission by Scott Vokes <vokes.s at gmail.com> (MAINTAINER)
with lots of tweaks by me and kurt@

help and ok kurt@