Import freedroidrpg-0.10.2
FreeDroidRPG is a free isometric RPG game inspired by elements of Diablo
and Fallout. Originally based on FreeDroid Classic, this project now has
vastly deviated from its parent.

This game tells the story of a world destroyed by a conflict between the
bots and their human masters. Play as Tux in a quest to save the world
from the murderous rebel bots who know no mercy...
... Or maybe just spend your time wandering around the land while the
people are slaughtered. You get to choose which path you wish to follow,
and freedom of choice is everywhere in the game.

FreeDroidRPG comes with an innovative system of magic, with features
such as forced casting and over 20 spells. You can use over 50 different
kinds of items and fight countless enemies on your way to your destiny.
We have an advanced dialogue system, with it's aim to be at least on par
with Fallout's. Our battle engine supports both ranged and close combat
in real time. Finally, if guns are too inaccurate and blades too messy,
you can always take over your enemies and have them fight on your side.


- joint work with FreeDroidRPG developer Arthur Huillet <ahuillet at freedroid dot org>
- thanks to Joachim Schipper <j.schipper at math.uu dot nl> for all the
feedbacks and testing
- disable on sparc64 ("Bus Error" on startup, I'll try to work on it)