Import celestia-1.4.1
Unlike most planetarium software, Celestia doesn't confine you to the
surface of the Earth. You can travel throughout the solar system, to any
of over 100,000 stars, or even beyond the galaxy.

All movement in Celestia is seamless; the exponential zoom feature lets
you explore space across a huge range of scales, from galaxy clusters
down to spacecraft only a few meters across. A 'point-and-goto'
interface makes it simple to navigate through the universe to the object
you want to visit.

Celestia is expandable. It comes with a large catalog of stars,
galaxies, planets, moons, asteroids, comets, and spacecraft. If that's
not enough, you can download dozens of easy to install add-ons with more
objects.

Available flavors:
	lua - add Lua scripting engine (extension language support)

ok jasper@