Initial import of Xinha 0.931
Xinha (pronounced like Xena, the Warrior Princess) is a powerful WYSIWYG
HTML editor component that works in Mozilla based browsers as well as in
MS Internet Explorer. Its configurabilty and extensibility make it easy to
build just the right editor for multiple purposes, from a restricted
mini-editor for one database field to a full-fledged website editor.