import xmmsctrl 1.9
xmmsctrl is a small utility to control xmms from the command line. Its
goal is to be used coupled with sh to test xmms state and perform an
appropriate action, e.g. if playing then pause else play. The interest
of this is to bind keys in a window manager to have control over xmms
with keys that do play/next/pause, prev, control sound.

by Gleydson Soares (MAINTAINER) with tweaks by me