Kill the phpxs script that was used for enabling and disabling the php core module and extenions. Install a sample configuration file to /var/www/conf/modules.sample which can be symlinked or copied over to /var/www/conf/modules so apache is going to pick it up.
Allow php to scan /var/www/conf/php5 for php configuration
files so if the user installs or creates a symlink from the
sample configuration files from the php5.sample directory,
it is going to be picked up by php5.

Create a dummy pwd.db file in the php5-imap package in the apache
chroot because it is needed by c-client.

feedback and tests by sthen@