import p5-Text-MultiMarkdown
Markdown is a text-to-HTML filter; it translates an easy-to-read /
easy-to-write structured text format into HTML.  Markdown's text format
is most similar to that of plain text email, and supports features such
as headers, *emphasis*, code blocks, blockquotes, and links.

This module implements MultiMarkdown, which is an extension to Markdown.

ok merdely@