import pessulus 2.16.4
Pessulus is a lockdown editor for GNOME, written in python. It enables
administrators to set mandatory settings in GConf. The users can not
change these settings.

Use of pessulus can be useful on computers that are open to use by
everyone, e.g. in an internet cafe.

ok ajacoutot@ (for a previous version of this port)