- tweak the way the gtk-update-icon-cache goo is run
gtk-update-icon-cache is part of gtk+2: adding gtk+2 to run_depends just
to update the icon cache (which only gtk apps can use) is overkill to
say the least!
As from now, each time icons are installed under %D/share/icons, we try
to execute gtk-update-icon-cache and if it is not there, we just ignore
the error.
What it means is that if you have gtk+2 installed, then it'll run fine
and your apps will be able to use the cache. Otherwise, it will silently
fails which is fine since it means none of your apps would have been
able to take advantage of the cache anyway.

discussed with jasper@