Import ekiga-2.0.12
Ekiga (formely known as GnomeMeeting) is an open source VoIP and video
conferencing application for GNOME.  Ekiga uses both the H.323 and SIP
protocols. It supports many audio and video codecs, and is interoperable
with other SIP compliant software and also with Microsoft NetMeeting.


It still needs some love but having it in-tree will ease people testing.
Some things are known not to be working yet...
Network part was intensively tested using the H.323 protocol through an
OpenBSD gnugk proxy, with Ekiga running OpenBSD/amd64 and a Polycom v500 as
endpoints.