import enigmail-seamonkey-0.95.6
Enigmail is an extension to the mail client of Mozilla/Netscape and
Mozilla Thunderbird which allows users to access the authentication
and encryption features provided by GnuPG.

This port is similar to enigmail, but is built for seamonkey.
Importing new port, instead of adding quirks to existing port, was
a much simplier solution this time.

regxpcom is an awk script, that takes care of maintaining
installed-chromes.txt.
go ahead, and a tweak from bernd@.  kurt@ agrees.