Import cheese-2.22.3
Cheese uses your webcam to take photos and videos, applies fancy special
effects and lets you share the fun with others.  Under the hood, Cheese
uses GStreamer to apply fancy effects to photos and videos.  With Cheese
it is easy to take photos of you, your friends, pets or whatever you
want and share them with others.

Add a patch from NetBSD to fix v4l2 support.

BEWARE!
This port is not functionnal right now, an update to
gstreamer-plugins-good is needed first