import p5-Statistics-Descriptive 2.6
This module provides basic functions used in descriptive statistics. It
has an object oriented design and supports two different types of data
storage and calculation objects: sparse and full. With the sparse
method, none of the data is stored and only a few statistical measures
are available. Using the full method, the entire data set is retained
and additional functions are available.