import Encode::Detect
This Perl module is an Encode::Encoding subclass that uses
Encode::Detect::Detector to determine the charset of the input data
and then decodes it using the encoder of the detected charset.

It is similar to Encode::Guess, but does not require the configuration
of a set of expected encodings. Like Encode::Guess, it only supports
decoding--it cannot encode.

ok simon@