import of mono-addins:
Mono.Addins is a framework for creating extensible applications,
and for creating libraries which extend those applications.
Mono.Addins has been designed to be easy to use and useful for
a wide range of applications: from simple applications with small
extensibility needs, to complex applications which need support
for large add-in structures.
This new framework intends to set an standard for building
extensible applications and add-ins in Mono.