import p5-CGI-Application-Plugin-DBH
CGI::Application::Plugin::DBH adds easy access to a DBI database handle
to your CGI::Application modules.  Lazy loading is used to prevent a
database connection from being made if the dbh method is not called
during the request.  In other words, the database connection is not
created until it is actually needed.

from Mikolaj Kucharski