import ircd-ratbox
ircd-ratbox is an advanced, stable and fast ircd.
It is the primary ircd used on EFNet, combining the stability of an ircd
required for a large production network together with a rich set of
features, making it also suitable for use on smaller networks. 

Features include:
 * optional SSL support to enable encrypted connections between clients
   and servers, as well as server to server links
 * support for SSL only channels (channel mode +S)
 * uses sqlite3 for handling and storing k/x/d lines
 * support for global CIDR limits
 * provides adminwall command allowing admins to broadcast messages to
   each other
 * support to force nick changes

request, testing and feedback by Michiel van Baak
setusercontext() idea from form@ via bernd@
tested on armish and ok by sthen@