Import genius-1.0.6
Genius is a general purpose calculator program similiar in some aspects 
to BC, Matlab, Maple or Mathematica.  It is useful both as a simple
calculator and as a research or educational tool.  The syntax is very
intuitive and is designed to mimic how mathematics is usually written.
GEL is the name of its extention language, it stands for Genius
Extension Language.