import discount, from maintainer Scott Vokes with minor tweaks:
This is David Loren Parsons's implementation of John Gruber's Markdown
text to html language. There's not much here that differentiates it from
any of the existing Markdown implementations except that it's written in
C instead of one of the vast flock of scripting languages that are
fighting it out for the Perl crown.

Markdown provides a library that gives you formatting functions suitable
for marking down entire documents or lines of text, a command-line
program that you can use to mark down documents interactively or from a
script, and a tiny suite of example programs that show how to fully
utilize the markdown library.