import libev 3.8, from MAINTAINER Baptiste Daroussin (with some tweaks by me):
Libev is a high-performance event loop/event model with lots of
features.  It is modelled (very loosely) after libevent and the Event
perl module, but aims to be faster and more correct, and also more
featureful. And also smaller.

(actually, really import the files. Oh yeah, cvs i love you so much.)