import net/pear-Net-DNS, from dlg@
A resolver library used to communicate with a name server to perform
DNS queries, zone transfers, dynamic DNS updates, etc.

Creates an object hierarchy from a DNS server response, which allows
you to view all of the information given by the DNS server. It
bypasses the system resolver library and communicates directly with
the server.