You are Agent Kate, one of the agents of the secret BlinkenArea special forces, also called the "BlinkenSisters". You have just been assigned to the following mission...
From the mission briefing:

An accident in the secret high-energy BlinkenArea lab has transfered
irreplaceable self-powered pixel hardware to various, random locations
in another dimension.

Your goal, Agent Kate, is to locate that hardware. We have fitted you
with an automated scanner/transmitter that will automatically transfer
the pixels back to the lab when you're close to it.

Beware that in dimension B15, while beeing still on earth, the scale of
things will differ from what you expect. Also, B15 is known to be home
to strange creatures that may pose a threat to you.

As soon as you have collected all pieces of hardware in one location,
you can be transfered to the next location.

Good luck, Agent Kate! 

---
looked at, and "its fun" from landry@ and jasper@