Update Wordpress to 2.8.6. From http://wordpress.org/development/2009/11/wordpress-2-8-6-security-release/:
2.8.6 fixes two security problems that can be exploited by registered,
logged in users who have posting privileges.  If you have untrusted
authors on your blog, upgrading to 2.8.6 is recommended.

The first problem is an XSS vulnerability in Press This discovered by
Benjamin Flesch.  The second problem, discovered by Dawid Golunski, is
an issue with sanitizing uploaded file names that can be exploited in
certain Apache configurations. Thanks to Benjamin and Dawid for finding
and reporting these.