backport important bug-fix from newest ltdl.
There are probably other bad bugs in there, this code is a textbook example
of how NOT to write code if you want it to work.

Between the gnu style that obfuscate the stuff, the various thingies added
"just in case", the misuse of autoconf results, and the really high
abstraction level together with long functions and internal concepts, how
can anyone make sense of this code ? (short answer: it doesn't make actual
sense).

Also, no cookies for the gnu guys who, along with an important bug-fix,
manage *again* to push out 10000+ lines of useless diff thanks to a change
in automake/autoconf itself.

We don't need to make fun of Windows  and its security, the GNU project
manages to produce as many problems on its own...

When will they learn secure practices ? I would bet "not this century",
but then I probably won't be around to collect the bet...