Re-import (from www). As noted by krw, it was present in Makefile but was not under net as it was supposed to be Original from landry@
Import p5-Net-SFTP-Foreign-1.55:

Net::SFTP::Foreign is a Perl client for the SFTP protocol version 3 as
defined in the SSH File Transfer Protocol IETF draft.
It uses any compatible ssh command installed on the system (for
instance, OpenSSH ssh) to establish the secure connection to the remote
server.
It is a lightweight alternative to p5-Net-SFTP.