import statusnet 0.8.2
StatusNet (formerly Laconica) is a Free and Open Source microblogging
platform. It helps people in a community, company or group to exchange
short (140 character) messages over the Web. Users can choose which
people to "follow" and receive only their friends' or colleagues' status
messages. It provides a similar service to sites like Twitter, Jaiku,
Yammer, and Plurk.

not hooking it up to the builds yet, as it will need some more tweaking
to set up.