- import p5-Device-USB 0.28
This module provides a Perl interface to the C library libusb. This
library supports a relatively full set of functionality to access a USB
device. In addition to the libusb, functioality, Device::USB provides a
few convenience features that are intended to produce a more Perl-ish
interface.

with help from landry@ to unfuck it's LIBS handling.