import sysutils/shmux, based on a submission from Olivier Cherrier long ago.  ok jasper@
shmux is program for executing the same command on many hosts in
parallel. For each target, a child process is spawned by shmux, and a
shell on the target obtained one of the supported methods: rsh, ssh, or
sh. The output produced by the children is received by shmux and either
(optionally) output in turn to the user using an easy to read format, or
written to files for later processing making it well suited for use in
scripts.