Import ports/devel/p5-Data-HexDump From maintainer Abel Abraham Camarillo Ojeda
Dump in hexadecimal the content of a scalar. The result is returned
in a string.  Each line of the result consists of the offset in the
source in the leftmost column of each line, followed by one or more
columns of data from the source in hexadecimal. The rightmost column
of each line shows the printable characters (all others are shown
as single dots).