import ports/graphics/darktable, thanks ajacoutot@ for review+fixes.
darktable is a virtual lighttable and darkroom for photographers:
it manages your digital negatives in a database and lets you view
them through a zoomable lighttable. It also enables you to develop
raw images and enhance them. It tries to fill the gap between the
many excellent existing free raw converters and image management
tools (such as ufraw or f-spot).

All editing is fully non-destructive and only operates on cached
image buffers for display. The full image is only converted during
export. Raw files, high-dynamic range and standard image formats
such as jpeg are all supported. The core operates completely on
floating point values, so darktable can not only be used for
photography but also for scientifically acquired images or output
of renderers (high dynamic range).