Remove MASTER_SITE_GITHUB
This was only used in two ports, and will never be used in more ports
as GitHub no longer builds new gems.  Both of the ports are also
found on rubygems.org, so this switches them to use the standard
MASTER_SITE_RUBYGEMS from ruby.port.mk. ruby-findmyiphone has a small
change from the GitHub version that requires a package bump.

OK landry@